5.14. Модуляция ДЗ
